requirements_resume_ats.txt
# Resume ATS Bot Requirements

# Core dependencies
streamlit>=1.28.0
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
chromadb

# Document processing
unstructured[docx,pdf]==0.16.16
PyPDF2>=3.0.1
pdfplumber>=0.11.0
python-docx>=1.1.0

# Firebase
firebase-admin>=6.5.0
google-cloud-storage>=2.14.0

# NLP (optional)
spacy>=3.7.0

# Utilities
python-dotenv>=1.0.0

# Numpy compatibility
numpy<2.0.0,>=1.22.4
